requirements.txt
websockets
aiohttp
requests
python-dotenv
orjson
web3
py-clob-client
py-builder-relayer-client
aiogram
filelock
uvloop
